Struct google_iam1::LogConfig

pub struct LogConfig {
    pub counter: Option<CounterOptions>,
    pub data_access: Option<DataAccessOptions>,
    pub cloud_audit: Option<CloudAuditOptions>,
}

Specifies what kind of log the caller must write Increment a streamz counter with the specified metric and field names. Metric names should start with a '/', generally be lowercase-only, and end in "_count". Field names should not contain an initial slash. The actual exported metric names will have "/iam/policy" prepended. Field names correspond to IAM request parameters and field values are their respective values. At present only "iam_principal", corresponding to IAMContext.principal, is supported. Examples: counter { metric: "/debug_access_count" field: "iam_principal" } ==> increment counter /iam/policy/backend_debug_access_count {iam_principal=[value of IAMContext.principal]} At this time we do not support: * multiple field names (though this may be supported in the future) * decrementing the counter * incrementing it by anything other than 1

This type is not used in any activity, and only used as part of another schema.

Fields
counter: Option<CounterOptions>

Counter options.

data_access: Option<DataAccessOptions>

Data access options.

cloud_audit: Option<CloudAuditOptions>

Cloud audit options.