Struct google_iam1::ProjectMethods

pub struct ProjectMethods<'a, C, A> where
    C: 'a,
    A: 'a,  { /* fields omitted */ }

A builder providing access to all methods supported on project resources. It is not used directly, but through the Iam hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate hyper_rustls;
extern crate yup_oauth2 as oauth2;
extern crate google_iam1 as iam1;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use iam1::Iam;
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::with_connector(hyper::net::HttpsConnector::new(hyper_rustls::TlsClient::new())),
                              <MemoryStorage as Default>::default(), None);
let mut hub = Iam::new(hyper::Client::with_connector(hyper::net::HttpsConnector::new(hyper_rustls::TlsClient::new())), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `roles_create(...)`, `roles_delete(...)`, `roles_get(...)`, `roles_list(...)`, `roles_patch(...)`, `roles_undelete(...)`, `service_accounts_create(...)`, `service_accounts_delete(...)`, `service_accounts_disable(...)`, `service_accounts_enable(...)`, `service_accounts_get(...)`, `service_accounts_get_iam_policy(...)`, `service_accounts_keys_create(...)`, `service_accounts_keys_delete(...)`, `service_accounts_keys_get(...)`, `service_accounts_keys_list(...)`, `service_accounts_list(...)`, `service_accounts_patch(...)`, `service_accounts_set_iam_policy(...)`, `service_accounts_sign_blob(...)`, `service_accounts_sign_jwt(...)`, `service_accounts_test_iam_permissions(...)`, `service_accounts_undelete(...)` and `service_accounts_update(...)`
// to build up your call.
let rb = hub.projects();

Methods

impl<'a, C, A> ProjectMethods<'a, C, A>

pub fn service_accounts_keys_delete(
&self,
name: &str
) -> ProjectServiceAccountKeyDeleteCall<'a, C, A>

Create a builder to help you perform the following task:

Deletes a ServiceAccountKey.

Arguments

name - The resource name of the service account key in the following format: projects/{PROJECT_ID}/serviceAccounts/{ACCOUNT}/keys/{key}. Using - as a wildcard for the PROJECT_ID will infer the project from the account. The ACCOUNT value can be the email address or the unique_id of the service account.

pub fn roles_get(&self, name: &str) -> ProjectRoleGetCall<'a, C, A>

Create a builder to help you perform the following task:

Gets a Role definition.

Arguments

name - The name parameter's value depends on the target resource for the request, namely roles, projects, or organizations. Each resource type's name value format is described below: * roles.get(): roles/{ROLE_NAME}. This method returns results from all predefined roles in Cloud IAM. Example request URL: https://iam.googleapis.com/v1/roles/{ROLE_NAME} * projects.roles.get(): projects/{PROJECT_ID}/roles/{CUSTOM_ROLE_ID}. This method returns only custom roles that have been created at the project level. Example request URL: https://iam.googleapis.com/v1/projects/{PROJECT_ID}/roles/{CUSTOM_ROLE_ID} * organizations.roles.get(): organizations/{ORGANIZATION_ID}/roles/{CUSTOM_ROLE_ID}. This method returns only custom roles that have been created at the organization level. Example request URL: https://iam.googleapis.com/v1/organizations/{ORGANIZATION_ID}/roles/{CUSTOM_ROLE_ID} Note: Wildcard (*) values are invalid; you must specify a complete project ID or organization ID.

pub fn service_accounts_enable(
&self,
request: EnableServiceAccountRequest,
name: &str
) -> ProjectServiceAccountEnableCall<'a, C, A>

Create a builder to help you perform the following task:

EnableServiceAccount is currently in the alpha launch stage.

Restores a disabled ServiceAccount that has been manually disabled by using DisableServiceAccount. Service accounts that have been disabled by other means or for other reasons, such as abuse, cannot be restored using this method.

EnableServiceAccount will have no effect on a service account that is not disabled. Enabling an already enabled service account will have no effect.

Arguments

request - No description provided.

name - The resource name of the service account in the following format: projects/{PROJECT_ID}/serviceAccounts/{ACCOUNT_UNIQUE_ID}'. Using -as a wildcard for thePROJECT_ID` will infer the project from the account.

pub fn service_accounts_get_iam_policy(
&self,
resource: &str
) -> ProjectServiceAccountGetIamPolicyCall<'a, C, A>

Create a builder to help you perform the following task:

Returns the Cloud IAM access control policy for a ServiceAccount.

Note: Service accounts are both resources and identities. This method treats the service account as a resource. It returns the Cloud IAM policy that reflects what members have access to the service account.

This method does not return what resources the service account has access to. To see if a service account has access to a resource, call the getIamPolicy method on the target resource. For example, to view grants for a project, call the projects.getIamPolicy method.

Arguments

resource - REQUIRED: The resource for which the policy is being requested. See the operation documentation for the appropriate value for this field.

pub fn roles_undelete(
&self,
request: UndeleteRoleRequest,
name: &str
) -> ProjectRoleUndeleteCall<'a, C, A>

Create a builder to help you perform the following task:

Undelete a Role, bringing it back in its previous state.

Arguments

request - No description provided.

name - The name parameter's value depends on the target resource for the request, namely projects or organizations. Each resource type's name value format is described below: * projects.roles.undelete(): projects/{PROJECT_ID}/roles/{CUSTOM_ROLE_ID}. This method undeletes only custom roles that have been created at the project level. Example request URL: https://iam.googleapis.com/v1/projects/{PROJECT_ID}/roles/{CUSTOM_ROLE_ID} * organizations.roles.undelete(): organizations/{ORGANIZATION_ID}/roles/{CUSTOM_ROLE_ID}. This method undeletes only custom roles that have been created at the organization level. Example request URL: https://iam.googleapis.com/v1/organizations/{ORGANIZATION_ID}/roles/{CUSTOM_ROLE_ID} Note: Wildcard (*) values are invalid; you must specify a complete project ID or organization ID.

pub fn roles_delete(&self, name: &str) -> ProjectRoleDeleteCall<'a, C, A>

Create a builder to help you perform the following task:

Soft deletes a role. The role is suspended and cannot be used to create new IAM Policy Bindings. The Role will not be included in ListRoles() unless show_deleted is set in the ListRolesRequest. The Role contains the deleted boolean set. Existing Bindings remains, but are inactive. The Role can be undeleted within 7 days. After 7 days the Role is deleted and all Bindings associated with the role are removed.

Arguments

name - The name parameter's value depends on the target resource for the request, namely projects or organizations. Each resource type's name value format is described below: * projects.roles.delete(): projects/{PROJECT_ID}/roles/{CUSTOM_ROLE_ID}. This method deletes only custom roles that have been created at the project level. Example request URL: https://iam.googleapis.com/v1/projects/{PROJECT_ID}/roles/{CUSTOM_ROLE_ID} * organizations.roles.delete(): organizations/{ORGANIZATION_ID}/roles/{CUSTOM_ROLE_ID}. This method deletes only custom roles that have been created at the organization level. Example request URL: https://iam.googleapis.com/v1/organizations/{ORGANIZATION_ID}/roles/{CUSTOM_ROLE_ID} Note: Wildcard (*) values are invalid; you must specify a complete project ID or organization ID.

pub fn service_accounts_test_iam_permissions(
&self,
request: TestIamPermissionsRequest,
resource: &str
) -> ProjectServiceAccountTestIamPermissionCall<'a, C, A>

Create a builder to help you perform the following task:

Tests the specified permissions against the IAM access control policy for a ServiceAccount.

Arguments

request - No description provided.

resource - REQUIRED: The resource for which the policy detail is being requested. See the operation documentation for the appropriate value for this field.

pub fn service_accounts_sign_blob(
&self,
request: SignBlobRequest,
name: &str
) -> ProjectServiceAccountSignBlobCall<'a, C, A>

Create a builder to help you perform the following task:

Note: This method is in the process of being deprecated. Call the signBlob() method of the Cloud IAM Service Account Credentials API instead.

Signs a blob using a service account's system-managed private key.

Arguments

request - No description provided.

name - The resource name of the service account in the following format: projects/{PROJECT_ID}/serviceAccounts/{ACCOUNT}. Using - as a wildcard for the PROJECT_ID will infer the project from the account. The ACCOUNT value can be the email address or the unique_id of the service account.

pub fn service_accounts_set_iam_policy(
&self,
request: SetIamPolicyRequest,
resource: &str
) -> ProjectServiceAccountSetIamPolicyCall<'a, C, A>

Create a builder to help you perform the following task:

Sets the Cloud IAM access control policy for a ServiceAccount.

Note: Service accounts are both resources and identities. This method treats the service account as a resource. Use it to grant members access to the service account, such as when they need to impersonate it.

This method does not grant the service account access to other resources, such as projects. To grant a service account access to resources, include the service account in the Cloud IAM policy for the desired resource, then call the appropriate setIamPolicy method on the target resource. For example, to grant a service account access to a project, call the projects.setIamPolicy method.

Arguments

request - No description provided.

resource - REQUIRED: The resource for which the policy is being specified. See the operation documentation for the appropriate value for this field.

pub fn service_accounts_delete(
&self,
name: &str
) -> ProjectServiceAccountDeleteCall<'a, C, A>

Create a builder to help you perform the following task:

Deletes a ServiceAccount.

Arguments

name - The resource name of the service account in the following format: projects/{PROJECT_ID}/serviceAccounts/{ACCOUNT}. Using - as a wildcard for the PROJECT_ID will infer the project from the account. The ACCOUNT value can be the email address or the unique_id of the service account.

pub fn service_accounts_create(
&self,
request: CreateServiceAccountRequest,
name: &str
) -> ProjectServiceAccountCreateCall<'a, C, A>

Create a builder to help you perform the following task:

Creates a ServiceAccount and returns it.

Arguments

request - No description provided.

name - Required. The resource name of the project associated with the service accounts, such as projects/my-project-123.

pub fn service_accounts_patch(
&self,
request: PatchServiceAccountRequest,
name: &str
) -> ProjectServiceAccountPatchCall<'a, C, A>

Create a builder to help you perform the following task:

Patches a ServiceAccount.

Currently, only the following fields are updatable: display_name and description.

Only fields specified in the request are guaranteed to be returned in the response. Other fields in the response may be empty.

Note: The field mask is required.

Arguments

request - No description provided.

name - The resource name of the service account in the following format: projects/{PROJECT_ID}/serviceAccounts/{ACCOUNT}. Requests using - as a wildcard for the PROJECT_ID will infer the project from the account and the ACCOUNT value can be the email address or the unique_id of the service account. In responses the resource name will always be in the format projects/{PROJECT_ID}/serviceAccounts/{ACCOUNT}.

pub fn roles_patch(
&self,
request: Role,
name: &str
) -> ProjectRolePatchCall<'a, C, A>

Create a builder to help you perform the following task:

Updates a Role definition.

Arguments

request - No description provided.

name - The name parameter's value depends on the target resource for the request, namely projects or organizations. Each resource type's name value format is described below: * projects.roles.patch(): projects/{PROJECT_ID}/roles/{CUSTOM_ROLE_ID}. This method updates only custom roles that have been created at the project level. Example request URL: https://iam.googleapis.com/v1/projects/{PROJECT_ID}/roles/{CUSTOM_ROLE_ID} * organizations.roles.patch(): organizations/{ORGANIZATION_ID}/roles/{CUSTOM_ROLE_ID}. This method updates only custom roles that have been created at the organization level. Example request URL: https://iam.googleapis.com/v1/organizations/{ORGANIZATION_ID}/roles/{CUSTOM_ROLE_ID} Note: Wildcard (*) values are invalid; you must specify a complete project ID or organization ID.

pub fn roles_create(
&self,
request: CreateRoleRequest,
parent: &str
) -> ProjectRoleCreateCall<'a, C, A>

Create a builder to help you perform the following task:

Creates a new Role.

Arguments

request - No description provided.

parent - The parent parameter's value depends on the target resource for the request, namely projects or organizations. Each resource type's parent value format is described below: * projects.roles.create(): projects/{PROJECT_ID}. This method creates project-level custom roles. Example request URL: https://iam.googleapis.com/v1/projects/{PROJECT_ID}/roles * organizations.roles.create(): organizations/{ORGANIZATION_ID}. This method creates organization-level custom roles. Example request URL: https://iam.googleapis.com/v1/organizations/{ORGANIZATION_ID}/roles Note: Wildcard (*) values are invalid; you must specify a complete project ID or organization ID.

pub fn service_accounts_disable(
&self,
request: DisableServiceAccountRequest,
name: &str
) -> ProjectServiceAccountDisableCall<'a, C, A>

Create a builder to help you perform the following task:

DisableServiceAccount is currently in the alpha launch stage.

Disables a ServiceAccount, which immediately prevents the service account from authenticating and gaining access to APIs.

Disabled service accounts can be safely restored by using EnableServiceAccount at any point. Deleted service accounts cannot be restored using this method.

Disabling a service account that is bound to VMs, Apps, Functions, or other jobs will cause those jobs to lose access to resources if they are using the disabled service account.

To improve reliability of your services and avoid unexpected outages, it is recommended to first disable a service account rather than delete it. After disabling the service account, wait at least 24 hours to verify there are no unintended consequences, and then delete the service account.

Arguments

request - No description provided.

name - The resource name of the service account in the following format: projects/{PROJECT_ID}/serviceAccounts/{ACCOUNT}. Using - as a wildcard for the PROJECT_ID will infer the project from the account. The ACCOUNT value can be the email address or the unique_id of the service account.

pub fn roles_list(&self, parent: &str) -> ProjectRoleListCall<'a, C, A>

Create a builder to help you perform the following task:

Lists the Roles defined on a resource.

Arguments

parent - The parent parameter's value depends on the target resource for the request, namely roles, projects, or organizations. Each resource type's parent value format is described below: * roles.list(): An empty string. This method doesn't require a resource; it simply returns all predefined roles in Cloud IAM. Example request URL: https://iam.googleapis.com/v1/roles * projects.roles.list(): projects/{PROJECT_ID}. This method lists all project-level custom roles. Example request URL: https://iam.googleapis.com/v1/projects/{PROJECT_ID}/roles * organizations.roles.list(): organizations/{ORGANIZATION_ID}. This method lists all organization-level custom roles. Example request URL: https://iam.googleapis.com/v1/organizations/{ORGANIZATION_ID}/roles Note: Wildcard (*) values are invalid; you must specify a complete project ID or organization ID.

pub fn service_accounts_undelete(
&self,
request: UndeleteServiceAccountRequest,
name: &str
) -> ProjectServiceAccountUndeleteCall<'a, C, A>

Create a builder to help you perform the following task:

Restores a deleted ServiceAccount. This is to be used as an action of last resort. A service account may not always be restorable.

Arguments

request - No description provided.

name - The resource name of the service account in the following format: projects/{PROJECT_ID}/serviceAccounts/{ACCOUNT_UNIQUE_ID}'. Using -as a wildcard for thePROJECT_ID` will infer the project from the account.

pub fn service_accounts_update(
&self,
request: ServiceAccount,
name: &str
) -> ProjectServiceAccountUpdateCall<'a, C, A>

Create a builder to help you perform the following task:

Note: This method is in the process of being deprecated. Use PatchServiceAccount instead.

Updates a ServiceAccount.

Currently, only the following fields are updatable: display_name and description.

Arguments

request - No description provided.

name - The resource name of the service account in the following format: projects/{PROJECT_ID}/serviceAccounts/{ACCOUNT}. Requests using - as a wildcard for the PROJECT_ID will infer the project from the account and the ACCOUNT value can be the email address or the unique_id of the service account. In responses the resource name will always be in the format projects/{PROJECT_ID}/serviceAccounts/{ACCOUNT}.

pub fn service_accounts_sign_jwt(
&self,
request: SignJwtRequest,
name: &str
) -> ProjectServiceAccountSignJwtCall<'a, C, A>

Create a builder to help you perform the following task:

Note: This method is in the process of being deprecated. Call the signJwt() method of the Cloud IAM Service Account Credentials API instead.

Signs a JWT using a service account's system-managed private key.

If no expiry time (exp) is provided in the SignJwtRequest, IAM sets an an expiry time of one hour by default. If you request an expiry time of more than one hour, the request will fail.

Arguments

request - No description provided.

name - The resource name of the service account in the following format: projects/{PROJECT_ID}/serviceAccounts/{ACCOUNT}. Using - as a wildcard for the PROJECT_ID will infer the project from the account. The ACCOUNT value can be the email address or the unique_id of the service account.

pub fn service_accounts_keys_list(
&self,
name: &str
) -> ProjectServiceAccountKeyListCall<'a, C, A>

Create a builder to help you perform the following task:

Lists ServiceAccountKeys.

Arguments

name - The resource name of the service account in the following format: projects/{PROJECT_ID}/serviceAccounts/{ACCOUNT}. Using - as a wildcard for the PROJECT_ID, will infer the project from the account. The ACCOUNT value can be the email address or the unique_id of the service account.

pub fn service_accounts_keys_create(
&self,
request: CreateServiceAccountKeyRequest,
name: &str
) -> ProjectServiceAccountKeyCreateCall<'a, C, A>

Create a builder to help you perform the following task:

Creates a ServiceAccountKey and returns it.

Arguments

request - No description provided.

name - The resource name of the service account in the following format: projects/{PROJECT_ID}/serviceAccounts/{ACCOUNT}. Using - as a wildcard for the PROJECT_ID will infer the project from the account. The ACCOUNT value can be the email address or the unique_id of the service account.

pub fn service_accounts_list(
&self,
name: &str
) -> ProjectServiceAccountListCall<'a, C, A>

Create a builder to help you perform the following task:

Lists ServiceAccounts for a project.

Arguments

name - Required. The resource name of the project associated with the service accounts, such as projects/my-project-123.

pub fn service_accounts_keys_get(
&self,
name: &str
) -> ProjectServiceAccountKeyGetCall<'a, C, A>

Create a builder to help you perform the following task:

Gets the ServiceAccountKey by key id.

Arguments

name - The resource name of the service account key in the following format: projects/{PROJECT_ID}/serviceAccounts/{ACCOUNT}/keys/{key}. Using - as a wildcard for the PROJECT_ID will infer the project from the account. The ACCOUNT value can be the email address or the unique_id of the service account.

pub fn service_accounts_get(
&self,
name: &str
) -> ProjectServiceAccountGetCall<'a, C, A>

Create a builder to help you perform the following task:

Gets a ServiceAccount.

Arguments

name - The resource name of the service account in the following format: projects/{PROJECT_ID}/serviceAccounts/{ACCOUNT}. Using - as a wildcard for the PROJECT_ID will infer the project from the account. The ACCOUNT value can be the email address or the unique_id of the service account.