Struct google_iam1::ProjectMethods

pub struct ProjectMethods<'a, C, A> where C: 'a, A: 'a {
    // some fields omitted
}

A builder providing access to all methods supported on project resources. It is not used directly, but through the Iam hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate yup_oauth2 as oauth2;
extern crate google_iam1 as iam1;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use iam1::Iam;
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = Iam::new(hyper::Client::new(), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `service_accounts_create(...)`, `service_accounts_delete(...)`, `service_accounts_get(...)`, `service_accounts_get_iam_policy(...)`, `service_accounts_keys_create(...)`, `service_accounts_keys_delete(...)`, `service_accounts_keys_get(...)`, `service_accounts_keys_list(...)`, `service_accounts_list(...)`, `service_accounts_set_iam_policy(...)`, `service_accounts_sign_blob(...)`, `service_accounts_test_iam_permissions(...)` and `service_accounts_update(...)`
// to build up your call.
let rb = hub.projects();

Methods

impl<'a, C, A> ProjectMethods<'a, C, A>

fn service_accounts_keys_list(&self, name: &str) -> ProjectServiceAccountKeyListCall<'a, C, A>

Create a builder to help you perform the following task:

Lists service account keys

Arguments

name - The resource name of the service account in the format "projects/{project}/serviceAccounts/{account}". Using '-' as a wildcard for the project, will infer the project from the account. The account value can be the email address or the unique_id of the service account.

fn service_accounts_set_iam_policy(&self, request: SetIamPolicyRequest, resource: &str) -> ProjectServiceAccountSetIamPolicyCall<'a, C, A>

Create a builder to help you perform the following task:

Sets the IAM access control policy for the specified IAM resource.

Arguments

request - No description provided.

resource - REQUIRED: The resource for which the policy is being specified. resource is usually specified as a path, such as projects/*project*/zones/*zone*/disks/*disk*. The format for the path specified in this value is resource specific and is specified in the setIamPolicy documentation.

fn service_accounts_keys_delete(&self, name: &str) -> ProjectServiceAccountKeyDeleteCall<'a, C, A>

Create a builder to help you perform the following task:

Deletes a service account key.

Arguments

name - The resource name of the service account key in the format "projects/{project}/serviceAccounts/{account}/keys/{key}". Using '-' as a wildcard for the project will infer the project from the account. The account value can be the email address or the unique_id of the service account.

fn service_accounts_keys_create(&self, request: CreateServiceAccountKeyRequest, name: &str) -> ProjectServiceAccountKeyCreateCall<'a, C, A>

Create a builder to help you perform the following task:

Creates a service account key and returns it.

Arguments

request - No description provided.

name - The resource name of the service account in the format "projects/{project}/serviceAccounts/{account}". Using '-' as a wildcard for the project, will infer the project from the account. The account value can be the email address or the unique_id of the service account.

fn service_accounts_sign_blob(&self, request: SignBlobRequest, name: &str) -> ProjectServiceAccountSignBlobCall<'a, C, A>

Create a builder to help you perform the following task:

Signs a blob using a service account.

Arguments

request - No description provided.

name - The resource name of the service account in the format "projects/{project}/serviceAccounts/{account}". Using '-' as a wildcard for the project, will infer the project from the account. The account value can be the email address or the unique_id of the service account.

fn service_accounts_update(&self, request: ServiceAccount, name: &str) -> ProjectServiceAccountUpdateCall<'a, C, A>

Create a builder to help you perform the following task:

Updates a service account. Currently, only the following fields are updatable: 'display_name' . The 'etag' is mandatory.

Arguments

request - No description provided.

name - The resource name of the service account in the format "projects/{project}/serviceAccounts/{account}". In requests using '-' as a wildcard for the project, will infer the project from the account and the account value can be the email address or the unique_id of the service account. In responses the resource name will always be in the format "projects/{project}/serviceAccounts/{email}".

fn service_accounts_list(&self, name: &str) -> ProjectServiceAccountListCall<'a, C, A>

Create a builder to help you perform the following task:

Lists service accounts for a project.

Arguments

name - Required. The resource name of the project associated with the service accounts, such as "projects/123"

fn service_accounts_delete(&self, name: &str) -> ProjectServiceAccountDeleteCall<'a, C, A>

Create a builder to help you perform the following task:

Deletes a service acount.

Arguments

name - The resource name of the service account in the format "projects/{project}/serviceAccounts/{account}". Using '-' as a wildcard for the project, will infer the project from the account. The account value can be the email address or the unique_id of the service account.

fn service_accounts_keys_get(&self, name: &str) -> ProjectServiceAccountKeyGetCall<'a, C, A>

Create a builder to help you perform the following task:

Gets the ServiceAccountKey by key id.

Arguments

name - The resource name of the service account key in the format "projects/{project}/serviceAccounts/{account}/keys/{key}". Using '-' as a wildcard for the project will infer the project from the account. The account value can be the email address or the unique_id of the service account.

fn service_accounts_get_iam_policy(&self, resource: &str) -> ProjectServiceAccountGetIamPolicyCall<'a, C, A>

Create a builder to help you perform the following task:

Returns the IAM access control policy for specified IAM resource.

Arguments

resource - REQUIRED: The resource for which the policy is being requested. resource is usually specified as a path, such as projects/*project*/zones/*zone*/disks/*disk*. The format for the path specified in this value is resource specific and is specified in the getIamPolicy documentation.

fn service_accounts_test_iam_permissions(&self, request: TestIamPermissionsRequest, resource: &str) -> ProjectServiceAccountTestIamPermissionCall<'a, C, A>

Create a builder to help you perform the following task:

Tests the specified permissions against the IAM access control policy for the specified IAM resource.

Arguments

request - No description provided.

resource - REQUIRED: The resource for which the policy detail is being requested. resource is usually specified as a path, such as projects/*project*/zones/*zone*/disks/*disk*. The format for the path specified in this value is resource specific and is specified in the testIamPermissions documentation.

fn service_accounts_create(&self, request: CreateServiceAccountRequest, name: &str) -> ProjectServiceAccountCreateCall<'a, C, A>

Create a builder to help you perform the following task:

Creates a service account and returns it.

Arguments

request - No description provided.

name - Required. The resource name of the project associated with the service accounts, such as "projects/123"

fn service_accounts_get(&self, name: &str) -> ProjectServiceAccountGetCall<'a, C, A>

Create a builder to help you perform the following task:

Gets a ServiceAccount

Arguments

name - The resource name of the service account in the format "projects/{project}/serviceAccounts/{account}". Using '-' as a wildcard for the project, will infer the project from the account. The account value can be the email address or the unique_id of the service account.