Struct google_iam1::ServiceAccount

pub struct ServiceAccount {
    pub oauth2_client_id: Option<String>,
    pub etag: Option<String>,
    pub display_name: Option<String>,
    pub name: Option<String>,
    pub unique_id: Option<String>,
    pub project_id: Option<String>,
    pub email: Option<String>,
}

A service account in the Identity and Access Management API. To create a service account, you specify the project_id and account_id for the account. The account_id is unique within the project, and used to generate the service account email address and a stable unique id. All other methods can identify accounts using the format "projects/{project}/serviceAccounts/{account}". Using '-' as a wildcard for the project, will infer the project from the account. The account value can be the email address or the unique_id of the service account.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

service accounts update projects (request|response)

service accounts get projects (response)

service accounts create projects (response)

Fields
oauth2_client_id: Option<String>

@OutputOnly. The OAuth2 client id for the service account. This is used in conjunction with the OAuth2 clientconfig API to make three legged OAuth2 (3LO) flows to access the data of Google users.

etag: Option<String>

Used to perform a consistent read-modify-write.

display_name: Option<String>

Optional. A user-specified description of the service account. Must be fewer than 100 UTF-8 bytes.

name: Option<String>

The resource name of the service account in the format "projects/{project}/serviceAccounts/{account}". In requests using '-' as a wildcard for the project, will infer the project from the account and the account value can be the email address or the unique_id of the service account. In responses the resource name will always be in the format "projects/{project}/serviceAccounts/{email}".

unique_id: Option<String>

@OutputOnly unique and stable id of the service account.

project_id: Option<String>

@OutputOnly The id of the project that owns the service account.

email: Option<String>

@OutputOnly Email address of the service account.